% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinstall.R
\name{checkinstall.plsda}
\alias{checkinstall.plsda}
\title{checkinstall.}
\usage{
checkinstall.plsda()
}
\value{
obj plsda object
}
\description{
functions to check if all packages are installed
checkinstall.plsda function to check package that the library pls needs
checkinstall.shiny.plsda function to check package that the shiny app pls needs
}
\examples{
checkinstall.plsda()
checkinstall.shiny.plsda()

}
