% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinstall.shiny.R
\name{checkinstall.shiny.plsda}
\alias{checkinstall.shiny.plsda}
\title{checkinstall.shiny.plsda}
\usage{
checkinstall.shiny.plsda()
}
\value{
obj plsda object
}
\description{
functions to check if all packages are installed
checkinstall.shiny.plsda function to check package that the shiny app pls needs
}
\examples{
checkinstall.shiny.plsda()

}
