% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export.plsda}
\alias{export.plsda}
\title{export.plsda}
\usage{
export.plsda(ypred)
}
\arguments{
\item{obj}{plsda object}
}
\value{
Value return
ypred prediction made by the plsda_predict function
}
\description{
function to export your prediction in a csv file
}
\examples{
obj = plsda()
obj = plsda_fit(obj,target~.,data,ncomp=2)
ypred = plsda_predict(obj, xtest,type = "posterior")
export.plsda(ypred)

}
