% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dummies.R
\name{get_dummies}
\alias{get_dummies}
\title{get_dummies}
\usage{
get_dummies(y)
}
\arguments{
\item{y}{a vector, factor that will be transformed, often used for your target variable}
}
\value{
dummies a matrix with all dummy values of the param in entry
}
\description{
This function allows you to convert categorical variables into numerical variables.
The different modalities are recorded in new columns that encode the presence of this category with binary values
}
\examples{
ydummies = get_dummies(iris$Species)

}
