% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{variableMap.plsda}
\title{plot}
\usage{
plot.scree(obj)
plot.varCorr(obj)
variableMap.plsda(obj,obj$x_scores$X1,obj$x_scores$X2)
}
\arguments{
\item{obj}{plsda object return by fit function}
}
\value{
plot plot to show information about data
}
\description{
plot.plsda : plots the variances against the number of the principal component.
plot.varCorr : show the correlation matrix in a pretty plot
variableMap.plsda : map of individuals on 2 composants
}
\examples{
obj = plsda()
obj =plsda_fit(obj,Species~., iris, ncomp=2)

plot.scree(obj)
plot.varCorr(obj)
variableMap.plsda(obj,obj$x_scores$X1,obj$x_scores$X2)

}
