% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scree.R
\name{plot.scree}
\alias{plot.scree}
\title{plot.scree()}
\usage{
plot.scree(obj)
}
\arguments{
\item{obj}{plsda object return by fit function}
}
\value{
plot plot to show eigen values of each composant
}
\description{
plot.scree : plots the variances against the number of the principal component.
}
\examples{
obj = plsda()
obj =plsda_fit(obj,Species~., iris, ncomp=2)

plot.scree(obj)

}
