% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.varCorr.R
\name{plot.varCorr}
\alias{plot.varCorr}
\title{plot.varCorr()}
\usage{
plot.varCorr(obj)
}
\arguments{
\item{obj}{plsda object return by fit function}
}
\value{
plot to show Correlation matrix
}
\description{
plot.varCorr : show the correlation matrix in a pretty plot
}
\examples{
obj = plsda()
obj =plsda_fit(obj,Species~., iris, ncomp=2)

plot.varCorr(obj)

}
