% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{plsda}
\alias{plsda}
\title{PLSDA Class}
\usage{
obj = plsda()
}
\arguments{
\item{ncomp}{number of composant}
}
\value{
obj plsda object
}
\description{
Class of the object PLSDA, this class is used to create an object Plsda, to use a Partial least squares regression method.
}
\examples{
obj = plsda()

objpls = plsda(ncomp=2)


print(obj)
summary(obj)

}
