% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_fit.R
\name{plsda_fit}
\alias{plsda_fit}
\title{plsda_fit}
\usage{
plsda_fit(obj,data$Species, data, ncomp=2)
}
\arguments{
\item{ObjectPLSDA}{PLDA Object
formula vector of the target data of the train dataset
data dataframe of the train dataset
ncomp integer number of composant
var.select boolean made a variable selection or not
centre boolean if the data need to be normalize}
}
\value{
obj plsda object
}
\description{
Method to use The Partial least squares regression
It initialize the method and train the model with the pls regression
to make a prediction on a dataset
}
\examples{
obj = plsda()
plsda_fit(obj,iris$Species, iris, ncomp=2)
}
