% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_predict.R
\name{plsda_predict}
\alias{plsda_predict}
\title{plsda_predict}
\usage{
plsda_predict(obj,newdata)
}
\arguments{
\item{obj}{PLDA object
newdata dataframe of the test dataset with no target variable
type "posterior" or "class" return format of the method}
}
\value{
type= posterior to have probability of class membership by observation
type= class to have the class predicted by observation
}
\description{
Method to use The Partial least squares regression Prediction
It make a prediction on test data
}
\examples{
obj = plsda()
obj = plsda_fit(obj,iris$Species,iris,ncomp=2)
newdata= iris[,-ncol(iris)]
plsda_predict(obj,newdata)



}
