% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_vip.R
\name{plsda_vip}
\alias{plsda_vip}
\title{Variable Importance in Projection}
\usage{
plsda_vip(Object, formula, threshold = 0.7)
}
\arguments{
\item{threshold}{threshold of variable's importance. 0.7 by default.}

\item{object}{A PLSDA class model.}
}
\value{
VIP : Most importants variables in the model.
\cr
newX : New variables
\cr
Newdataset : New dataset constructed with the selected variables
}
\description{
This function search for best variables to construct the best model.
}
\examples{
plsda_vip=(object = objPLSDA)
plsda_vip=(object = objPLSDA,threshold=0.8)

}
