% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{print.plsda}
\alias{print.plsda}
\title{print function for PLSDA Class}
\usage{
obj = plsda()
}
\arguments{
\item{obj}{an pls object}
}
\value{
observe value of an pls object
}
\description{
Overload of print function
}
\examples{
obj = plsda()
print(obj)
}
