% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{summary.plsda}
\alias{summary.plsda}
\title{summary for PLSDA Class}
\usage{
summary(obj)
}
\arguments{
\item{obj}{an pls object}
}
\value{
observe value of explicative varaibles in an object pls
}
\description{
Overload of summary function
}
\examples{
obj = plsda()
summary(obj)

}
