% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableMap.plsda.R
\name{variableMap.plsda}
\alias{variableMap.plsda}
\title{variableMap.plsda()}
\usage{
variableMap.plsda(obj)
}
\arguments{
\item{obj}{plsda object return by fit function}
}
\value{
plot plot to show individus projection in a new plan with 2 composants
}
\description{
variableMap.plsda : map of individuals on 2 composants
}
\examples{
obj = plsda()
obj =plsda_fit(obj,Species~., iris, ncomp=2)
variableMap.plsda(obj,obj$x_scores$X1,obj$x_scores$X2)

}
