% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinstall.R
\name{checkinstall.plsda}
\alias{checkinstall.plsda}
\title{checkinstall.plsda}
\usage{
checkinstall.plsda()
}
\description{
functions to check if all packages are installed
checkinstall.plsda function to check package that the library pls needs
}
\examples{
checkinstall.plsda()

}
