% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export.plsda}
\alias{export.plsda}
\title{export.plsda}
\usage{
export.plsda(ypred)
}
\arguments{
\item{obj}{plsda object
ypred prediction made by the plsda_predict function
path localisation of the file in your local machine. Be careful to use backslash between directory and give a name to your file at the end of the path}
}
\value{
prediction in a csv file
}
\description{
function to export your prediction in a csv file
}
\examples{
obj = plsda()
obj = plsda_fit(obj,target~.,data,ncomp=2)
ypred = plsda_predict(obj, xtest,type = "posterior")

export.plsda(ypred)
export.plsda(iris,"C:/Users/Maartin/Downloads/NameoftheFilePrediction.csv")

}
