% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.plsda}
\title{plot}
\usage{
plot.scree(obj)
plot.varCorr(obj)
}
\arguments{
\item{obj}{plsda object return by fit function}
}
\value{
plot plot to show informtions about data
}
\description{
plot.plsda : plots the variances against the number of the principal component.
plot.varCorr : show the correlation matrix in a pretty plot
}
\examples{
obj = plsda()
plsda_fit(obj,Species~., iris, ncomp=2)

plot.scree(obj)
plot.varCorr(obj)

}
